# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.Image:
    properties:
      url:
        type: string
    type: object
...
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.PreProduct:
    description: safe_deal optional delivery optional
    properties:
      available_count:
        type: integer
      category_id:
        type: integer
      city_id:
        type: integer
      delivery:
        type: boolean
      description:
        description: nolint
        type: string
      images:
        items:
          $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.Image'
        type: array
      is_active:
        type: boolean
      price:
        type: integer
      safe_deal:
        type: boolean
      saler_id:
        type: integer
      title:
        type: string
    type: object
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.Product:
    properties:
      available_count:
        type: integer
      category_id:
        type: integer
      city_id:
        type: integer
      created_at:
        type: string
      delivery:
        type: boolean
      description:
        description: nolint
        type: string
      favourites:
        type: integer
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.Image'
        type: array
      in_favourites:
        type: boolean
      is_active:
        type: boolean
      price:
        type: integer
      safe_deal:
        type: boolean
      saler_id:
        type: integer
      title:
        type: string
      views:
        type: integer
    type: object
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.ProductInFeed:
    properties:
      available_count:
        type: integer
      city_id:
        type: integer
      delivery:
        type: boolean
      favourites:
        type: integer
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.Image'
        type: array
      in_favourites:
        type: boolean
      is_active:
        type: boolean
      price:
        type: integer
      safe_deal:
        type: boolean
      title:
        type: string
    type: object
...
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse:
    properties:
      body:
        $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseBodyError'
      status:
        type: integer
    type: object
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseBody:
    properties:
      message:
        type: string
    type: object
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseBodyError:
    properties:
      error:
        type: string
    type: object
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseBodyID:
    properties:
      id:
        type: integer
    type: object
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseID:
    properties:
      body:
        $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseBodyID'
      status:
        type: integer
    type: object
  github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseSuccessful:
    properties:
      body:
        $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseBody'
      status:
        type: integer
    type: object
...
  internal_pkg_product_delivery.ProductListResponse:
    properties:
      body:
        items:
          $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.ProductInFeed'
        type: array
      status:
        type: integer
    type: object
  internal_pkg_product_delivery.ProductResponse:
    properties:
      body:
        $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.Product'
      status:
        type: integer
    type: object
...
host: 84.23.53.28:8080
info:
  contact: {}
  description: This is a server of YULA server.
  title: YULA project API
  version: "1.0"
paths:
...
  /product/activate:
    patch:
      consumes:
      - application/json
      description: |-
        activate product for saler using user id from cookies\jwt.
        This does product active.
      parameters:
      - description: product id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseSuccessful'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: activate product
      tags:
      - product
  /product/add:
    post:
      consumes:
      - application/json
      description: |-
        add product by data
        Error.status can be:
        StatusErrBadRequest      = 400
        StatusErrInternalServer  = 500
      parameters:
      - description: product data for adding
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.PreProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseID'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: add product
      tags:
      - product
...
  /product/close:
    patch:
      consumes:
      - application/json
      description: |-
        close product for saler using user id from cookies\jwt.
        This does product not active.
      parameters:
      - description: product id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseSuccessful'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: close product
      tags:
      - product
  /product/delete:
    delete:
      consumes:
      - application/json
      description: |-
        delete product for saler using user id from cookies\jwt.
        This totally removed product. Recovery will be impossible
      parameters:
      - description: product id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseSuccessful'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: delete product
      tags:
      - product
  /product/get:
    get:
      consumes:
      - application/json
      description: get product by id
      parameters:
      - description: product id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_pkg_product_delivery.ProductResponse'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get product
      tags:
      - product
  /product/get_list:
    get:
      consumes:
      - application/json
      description: get products by count and last_id return old products
      parameters:
      - description: count products
        in: query
        name: count
        required: true
        type: integer
      - description: last product id
        in: query
        name: last_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_pkg_product_delivery.ProductListResponse'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get products list
      tags:
      - product
  /product/get_list_of_another_saler:
    get:
      consumes:
      - application/json
      description: get list of products for another saler using saler id, count and
        last product id from query
      parameters:
      - description: saler id
        in: query
        name: saler_id
        required: true
        type: integer
      - description: count products
        in: query
        name: count
        required: true
        type: integer
      - description: 'last product id '
        in: query
        name: last_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_pkg_product_delivery.ProductListResponse'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get list of products for another saler
      tags:
      - product
  /product/get_list_of_saler:
    get:
      consumes:
      - application/json
      description: get list of products for saler using user id from cookies\jwt
      parameters:
      - description: count products
        in: query
        name: count
        required: true
        type: integer
      - description: 'last product id '
        in: query
        name: last_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_pkg_product_delivery.ProductListResponse'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get list of products for saler
      tags:
      - product
...
  /product/update:
    patch:
      consumes:
      - application/json
      description: update product by id
      parameters:
      - description: product id
        in: query
        name: id
        required: true
        type: integer
      - description: полностью опционален
        in: body
        name: preProduct
        schema:
          $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.PreProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseID'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: update product
      tags:
      - product
    put:
      consumes:
      - application/json
      description: update product by id
      parameters:
      - description: product id
        in: query
        name: id
        required: true
        type: integer
      - description: полностью опционален
        in: body
        name: preProduct
        schema:
          $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_models.PreProduct'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ResponseID'
        "222":
          description: Error
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2023_2_Rabotyagi_internal_pkg_server_delivery.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: update product
      tags:
      - product
...
schemes:
- http
swagger: "2.0"